requirements.txt
django==4.2.7
python-dotenv==1.0.0
PyMuPDF==1.23.7
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-google-genai==0.0.5
langchain-groq==0.1.4
google-generativeai==0.3.2
faiss-cpu==1.7.4